+++ routes/users.js
@@ -1,10 +1,9 @@
 const express = require("express");
 const {users} = require("../data/users.json");
 const { UserModel, BookModel } = require("../modals");
-const { getAllUsers, getSingleUserById, deleteUser } = require("../controllers/user-controller");
+const { getAllUsers, getSingleUserById, deleteUser, updateUserById, createNewUser, getSubscriptionDetailsById } = require("../controllers/user-controller");
 
 const router = express.Router();
-
 
 /**
  * Route: /users
@@ -31,30 +30,7 @@ router.get('/:id', getSingleUserById);
  * Access: Public
  * Paramaters: None
  */
-router.post('/', (req, res)=>{
-    const {id, name, surname, email, subscriptionType, subscriptionDate} = req.body;
-
-    const user = users.find((each)=> each.id === id);
-
-    if(user){
-        return res.status(404).json({
-            success: false,
-            message: "User With the Id already exists"
-        })
-    }
-    users.push({
-        id,
-        name, 
-        surname,
-        email,
-        subscriptionType,
-        subscriptionDate
-    })
-    return res.status(201).json({
-        success: true,
-        data: users
-    })
-})
+router.post('/', createNewUser);
 
 /**
  * Route: /users/:id
@@ -63,29 +39,7 @@ router.post('/', (req, res)=>{
  * Access: Public
  * Paramaters: ID
  */
-router.put('/:id', (req, res)=>{
-    const {id} = req.params;
-    const {data} = req.body;
-
-    const user = users.find((each)=> each.id === id);
-
-    if(!user)
-    return res.status(404).json({success: false, message: "User Does Not Exist"});
-
-    const updatedUser = users.map((each)=>{
-        if(each.id===id){
-            return {
-                ...each,
-                ...data,
-            }
-        }
-        return each;
-    })
-    return res.status(200).json({
-        success: true,
-        data: updatedUser
-    })
-})
+router.put('/:id', updateUserById)
 
 /**
  * Route: /users/:id
@@ -96,6 +50,7 @@ router.put('/:id', (req, res)=>{
  */
 router.delete('/:id', deleteUser);
 
+
 /**
  * Route: /users/subscription-details/:id
  * Method: GET
@@ -103,56 +58,7 @@ router.delete('/:id', deleteUser);
  * Access: Public
  * Paramaters: ID
  */
-router.get('/subscription-details/:id', (req, res)=>{
-    const {id} = req.params;
-
-    const user = users.find((each)=> each.id === id);
-    if(!user)
-        return res.status(404).json({success: false, message: "User With The Given Id Doesn't Exist"});
-
-    const getDateInDays = (data = "")=> {
-        let date;
-        if(data === ""){
-            // current Date
-            date = new Date();
-        }else{
-            // getting date on a basis of data variable
-            date = new Date(data);
-        }
-        let days = Math.floor(data / (1000 * 60 * 60 * 24));
-        return days;
-    };
-
-    const subscriptionType = (date) => {
-        if(user.subscriptionType === "Basic"){
-            date = date + 90;
-        }else if(user.subscriptionType === "Standard"){
-            date = date + 180;
-        }else if(user.subscriptionType === "Premium"){
-            date = date + 365;
-        }
-        return date;
-    };
-
-    // Subscription expiration calcus
-    // Jan 1, 1970, UTC // milliseconds
-    let returnDate = getDateInDays(user.returnDate);
-    let currentDate = getDateInDays();
-    let subscriptionDate = getDateInDays(user.subscriptionDate);
-    let subscriptionExpiration = subscriptionType(subscriptionDate);
-
-    const data = {
-        ...user,
-        subscriptionExpired: subscriptionExpiration < currentDate,
-        daysLeftForExpiration: subscriptionExpiration <= currentDate ? 0 : subscriptionDate - currentDate,
-        fine: returnDate < currentDate ? subscriptionExpiration <= currentDate ? 200 : 100 : 0, 
-    }
-
-     res.status(200).json({
-        success: true,
-        data: data,
-     })
-})
+router.get('/subscription-details/:id', getSubscriptionDetailsById);
 
 // default export
 module.exports = router;
